requirements.txt
kafka-python==2.0.2
tabulate==0.9.0
pyspark==3.4.1
cassandra-driver==3.28.0

